arrange internal grpc server methods
